gemini_grounded/requirements.txt
markdown2
gradio
google-generativeai